nginx-ssl-util: pull out common
