+++ net/nginx-utils/src/common.hpp
@@ -5,9 +5,29 @@
 #include <iostream>
 #include <string>
 #include <fstream>
+#include <unistd.h>
+
+
+void write_file(const std::string & filename, const std::string str,
+                std::ios_base::openmode flag=std::ios::trunc);
 
 
 std::string read_file(const std::string & filename);
+
+
+int call(const char program[], const char arg[]);
+
+
+
+void write_file(const std::string & filename, const std::string str,
+                std::ios_base::openmode flag)
+{
+    std::ofstream file (filename, std::ios::out|flag);
+    if (file.good()) { file<<str<<std::endl; }
+    file.close();
+}
+
+
 std::string read_file(const std::string & filename)
 {
     std::ifstream file(filename, std::ios::in|std::ios::ate);
@@ -23,14 +43,20 @@ std::string read_file(const std::string & filename)
     return str;
 }
 
-void write_file(const std::string & filename, const std::string str,
-                std::ios_base::openmode flag=std::ios::trunc);
-void write_file(const std::string & filename, const std::string str,
-                std::ios_base::openmode flag)
+
+int call(const char program[], const char arg[])
 {
-    std::ofstream file (filename, std::ios::out|flag);
-    if (file.good()) { file<<str<<std::endl; }
-    file.close();
+    pid_t pid = fork();
+    switch(pid) {
+        case -1: // could not fork.
+            return -1;
+        case 0: // child, exec never returns.
+            execl(program, program, arg, (char *)NULL);
+            exit(EXIT_FAILURE);
+        default: //parent
+            return pid;
+    }
 }
+
 
 #endif
